push19.02

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -89,6 +89,9 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::find($id);
+        if($category->posts->count()){
+            return redirect()->route('categories.index')->with('error',"Ошибка! У категории есть записи.");
+        };
         $category->delete();
         return redirect()->route('categories.index')->with('success',"Категория id:{$category->id}  \"{$category->title}\" удалена");
     }
