+++ app/Http/Controllers/Admin/TagController.php
@@ -15,8 +15,8 @@ class TagController extends Controller
      */
     public function index()
     {
-        $tags=Tag::paginate(10);
-        return view('admin.tags.index',compact('tags'));
+        $tags = Tag::paginate(10);
+        return view('admin.tags.index', compact('tags'));
     }
 
     /**
@@ -33,66 +33,68 @@ class TagController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $request->validate([
-            'title'=>'required',
+            'title' => 'required',
         ]);
 
         Tag::create($request->all());
         //$request->session()->flash('success','Категория добавлена');
-        return redirect()->route('tags.index')->with('success','Категория добавлена');
+        return redirect()->route('tags.index')->with('success', 'Категория добавлена');
     }
-
 
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
         $tag = Tag::find($id);
 
-        return view('admin.tags.edit',compact('tag'));
+        return view('admin.tags.edit', compact('tag'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
 
         $request->validate([
-            'title'=>'required',
+            'title' => 'required',
         ]);
 
         $tag = Tag::find($id);
-        $oldTag=$tag->title;
+        $oldTag = $tag->title;
 //        $tag->slug = null;
         $tag->update($request->all());
-        return redirect()->route('tags.index')->with('success',"Тэг id:{$tag->id} \"{$oldTag}\" переименован в \"{$tag->title}\" ");
+        return redirect()->route('tags.index')->with('success', "Тэг id:{$tag->id} \"{$oldTag}\" переименован в \"{$tag->title}\" ");
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $tag = Tag::find($id);
+        if ($tag->posts->count()) {
+            return redirect()->route('tags.index')->with('error', "Ошибка ! У тегов есть посты");
+        }
         $tag->delete();
-        return redirect()->route('tags.index')->with('success',"Тэг id:{$tag->id}  \"{$tag->title}\" удален");
+        return redirect()->route('tags.index')->with('success', "Тэг id:{$tag->id}  \"{$tag->title}\" удален");
     }
 }
